pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "memory-system-mcp"
version = "2.0.0"
description = "AI Memory System with MCP Server - Multi-persona support, skill-driven self-evolution"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Memory System Team" }
]
keywords = ["mcp", "memory", "ai", "claude", "llm", "skill-system"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "python-dateutil>=2.8.0",
    "chromadb>=0.4.0",
    "zhipuai>=2.0.0",
    "openai>=1.0.0",
    "requests>=2.28.0",
    "jieba>=0.42.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
memory-system-mcp = "memory_system.mcp_server:main"
memory-system-daemon = "memory_system.skill_evolution_daemon.service:main"

[project.urls]
Homepage = "https://github.com/ab123-ch/memory-system"
Documentation = "https://github.com/ab123-ch/memory-system#readme"
Repository = "https://github.com/ab123-ch/memory-system"

[tool.hatch.build.targets.wheel]
packages = ["memory_system"]

[tool.hatch.build.targets.sdist]
include = [
    "/memory_system",
    "/memory_mcp_server.py",
    "/memory_config.yaml",
    "/model_config.yaml",
    "/templates",
]
